S1: water core -> int32 Q16.16 (first fixed-point field migration)
Convert the SYNCED shallow-water state — water_depth (CONSERVED), flow_vx,
flow_vy, floor_height — from float32 to int32 Q16.16 (scale 2^16). The
donor-cell upwind transport core is now pure integer (bit-identical
cross-machine); the ripple stays float (render-only) reading dequantized depth.

Toolkit: cpp/src/fixed_point.h — q16 alias, mul_q16 (truncating >>16, matching
temperature_solver), quantize (round-to-nearest), recip_mul (128-bit reciprocal
multiply, MSVC _mul128 path), shr_round0, ceil_div (the substep cliff), tan_poly
(degree-5 odd Taylor). Python mirror: src/simulation/water_fixed.py.

Conversion:
  * surface potential: floor + tilt + depth in Q16.16; tilt via tan_poly on the
    clamped (+/-35deg) scalar tilt.
  * velocity kick: g*dt / damp*dt as Q16.16 step constants; central-difference
    grad via recip_mul of 2*dx.
  * flux: gathered ONCE as int64 mul_wide(v_face, depth_donor); per-FACE depth-
    delta dq narrowed once and applied +dq/-dq to the two cells -> mass conserved
    to the LSB at the >>16 narrow (P2).
  * outflow limiter: dynamic depth/out_sum as an exact integer divide (the dx/dt
    constant folded into dq).
  * substep-count cliff: n = ceil_div(quantize(sim_time), max_dt_q()) — integer,
    bit-identical across peers. max_dt's sqrt computed once in double at load.

FLOAT BRIDGES until S2 (atmosphere/smoke/gas still float): the k_p head term,
the W5 boil (dequantize depth, boil in integer counts so water removed == steam
credited), and the W3 P*V displacement + flooded seal — each dequantizes depth
at the boundary, marked in-source.

Gates: P2 Σwater_depth bit-conserved (0 LSB drift over 1000 ticks, incl. tilt
and limiter-stress checkerboard at 3x CFL); P1 bit-identity (field A/B harness,
per-cell 0-ULP); feel-regression vs float baseline ~sub-mm depth, ~cm/s flow;
350 tests pass; both --auto exit 0. no-float ratchet re-baselined (water core
int; doubles are load-time constants, the 4 W5/W3 floats are bridges).

+++ cpp/src/physics_engine.h
@@ -65,7 +65,7 @@ public:
     std::vector<std::pair<int, int>> step_tail(
         // ripple group
         float* ripple, float* ripple_v,
-        const float* water_depth, const float* wave_p,
+        const int32_t* water_depth, const float* wave_p,   // S1: water_depth Q16.16
         const bool* solid,
         // fire group
         float* fire_field, float* atmosphere, float* smoke_field, float* wall_hp,
@@ -173,12 +173,16 @@ public:
     //   tilt_x, tilt_y                : solver tilt (read)
     //   sim_time                      : tick length (seconds)
     //   ceiling_h..steam_yield        : the W3/W5 scalar params (Python doubles)
+    // S1: water_depth/flow_vx/flow_vy/floor_height/before are now int32 Q16.16
+    // (metres / m/s). atmosphere/gas/dyn_permeability stay FLOAT (the S2 group) —
+    // the W5 boil + W3 displacement are FLOAT BRIDGES that dequantize water_depth
+    // at the boundary (marked in the .cpp). The substep-count cliff is integer.
     void step_water(
-        float* water_depth, float* flow_vx, float* flow_vy,
-        const float* floor_height, float* atmosphere, const float* wave_p,
+        int32_t* water_depth, int32_t* flow_vx, int32_t* flow_vy,
+        const int32_t* floor_height, float* atmosphere, const float* wave_p,
         const bool* solid,
         float* gas,
-        float* before, float* dyn_permeability,
+        int32_t* before, float* dyn_permeability,
         int steam_idx, float tilt_x, float tilt_y,
         int h, int w, float sim_time,
         double ceiling_h, double flood_eps, double ratio_cap,
